+++ lib/AccessControl/Form/FormAccessControl/FormAccessControl.js
@@ -7,9 +7,9 @@ import { FormattedMessage } from 'react-intl';
 
 import { Accordion, Badge, Spinner } from '@folio/stripes/components';
 
-import { useClaimPolicies, useDoAccessControl } from '../hooks';
-import { PolicyTypedown } from './PolicyTypedown';
-import { PoliciesTable } from '../View';
+import { useClaimPolicies, useDoAccessControl } from '../../hooks';
+import { PolicyTypedown } from '../PolicyTypedown';
+import { PoliciesTable } from '../../View';
 
 const FormAccessControl = ({
   accessControlEndpoint, // THIS MUST BE THE BASE endpoint for accessControl
