+++ lib/AccessControl/Form/FormAccessControl/FormAccessControl.test.js
@@ -16,13 +16,15 @@ import {
   TestForm
 } from '@folio/stripes-erm-testing';
 
-import { useDoAccessControl } from '../hooks';
+import { useDoAccessControl } from '../../hooks';
 import FormAccessControl from './FormAccessControl';
 
-import { translationsProperties } from '../../../test/jest/helpers';
+import { translationsProperties } from '../../../../test/jest/helpers';
 
+// Ensure we have the centralised claim policies
+jest.mock('../../hooks/useClaimPolicies');
 
-jest.mock('../hooks/useDoAccessControl', () => jest.fn(() => ({
+jest.mock('../../hooks/useDoAccessControl', () => jest.fn(() => ({
   doAccessControl: true,
   enabledEnginesQuery: {
     isLoading: false
@@ -32,12 +34,9 @@ jest.mock('../hooks/useDoAccessControl', () => jest.fn(() => ({
 // We should no longer need to mock icons -- needed for Spinner check
 jest.unmock('@folio/stripes-components');
 
-// Ensure we have the centralised claim policies
-jest.mock('../hooks/useClaimPolicies');
-
 
 // Mock PolicyTypedown, we don't need the full component here
-jest.mock('./PolicyTypedown', () => {
+jest.mock('../PolicyTypedown', () => {
   // Mock the PolicyTypedown with the ability to select policies and see if they're selected
   const MockPolicyTypedown = ({ input: { onChange } = {}, isSelected, policies }) => {
     return (
@@ -68,7 +67,7 @@ jest.mock('./PolicyTypedown', () => {
   });
 });
 
-jest.mock('../View', () => {
+jest.mock('../../View', () => {
   // Mock the PoliciesTable to return JUST a list of selected policy names
   const MockPoliciesTable = ({ policies }) => {
     return (
