+++ lib/AccessControl/Form/FormAccessControl/index.js
@@ -1,0 +1,1 @@
+export { default } from './FormAccessControl';
