+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/AcquisitionUnitPolicy/AcquisitionUnitPolicy.js
@@ -3,12 +3,10 @@ import { useIntl } from 'react-intl';
 import { Layout } from '@folio/stripes/components';
 import { highlightString } from '@k-int/stripes-kint-components';
 
-import Separator from './Separator';
-import css from '../Styles.css';
+import { acquisitionPolicyRestrictions } from '../acquisitionPolicyRestrictions';
 
-export const acquisitionPolicyRestrictions = (policy, intl) => ['protectRead', 'protectUpdate', 'protectCreate', 'protectDelete'].map(restriction => {
-  return policy[restriction] ? intl.formatMessage({ id: `stripes-erm-components.accesscontrol.acquisitionunits.${restriction}` }) : null;
-}).filter(Boolean).join(', ');
+import Separator from '../Separator';
+import css from '../../Styles.css';
 
 const AcquisitionUnitPolicy = ({
   isSelected,
