+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/AcquisitionUnitPolicy/AcquisitionUnitPolicy.test.js
@@ -2,11 +2,11 @@ import { within } from '@folio/jest-config-stripes/testing-library/react';
 
 import { renderWithIntl } from '@folio/stripes-erm-testing';
 import AcquisitionUnitPolicy from './AcquisitionUnitPolicy';
-import { translationsProperties } from '../../../../../test/jest/helpers';
-import { useClaimPolicies } from '../../../hooks';
+import { translationsProperties } from '../../../../../../test/jest/helpers';
+import { useClaimPolicies } from '../../../../hooks';
 
 // Use manual mock which can get these properly
-jest.mock('../../../hooks/useClaimPolicies');
+jest.mock('../../../../hooks/useClaimPolicies');
 
 const TestComponent = ({ typed, isSelected }) => {
   const { flattenedClaimPolicies } = useClaimPolicies({});
