+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/AcquisitionUnitPolicy/index.js
@@ -1,0 +1,1 @@
+export { default } from './AcquisitionUnitPolicy';
