+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/Separator/Separator.test.js
@@ -1,7 +1,7 @@
 import { renderWithIntl } from '@folio/stripes-erm-testing';
 
 import Separator from './Separator';
-import { translationsProperties } from '../../../../../test/jest/helpers';
+import { translationsProperties } from '../../../../../../test/jest/helpers';
 
 let renderComponent;
 describe('Separator', () => {
