+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/Separator/index.js
@@ -1,0 +1,1 @@
+export { default } from './Separator';
