+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/acquisitionPolicyRestrictions.js
@@ -1,0 +1,3 @@
+export const acquisitionPolicyRestrictions = (policy, intl) => ['protectRead', 'protectUpdate', 'protectCreate', 'protectDelete'].map(restriction => {
+  return policy[restriction] ? intl.formatMessage({ id: `stripes-erm-components.accesscontrol.acquisitionunits.${restriction}` }) : null;
+}).filter(Boolean).join(', ');
