+++ lib/AccessControl/Form/PolicyTypedown/PolicyRenderComponents/index.js
@@ -1,2 +1,3 @@
 export { default as Separator } from './Separator';
-export { default as AcquisitionUnitPolicy, acquisitionPolicyRestrictions } from './AcquisitionUnitPolicy';
+export { default as AcquisitionUnitPolicy } from './AcquisitionUnitPolicy';
+export { acquisitionPolicyRestrictions } from './acquisitionPolicyRestrictions';
