+++ lib/AccessControl/Form/PolicyTypedown/PolicyTypedown/PolicyTypedown.js
@@ -5,8 +5,8 @@ import isEqual from 'lodash/isEqual';
 
 import { Typedown } from '@k-int/stripes-kint-components';
 
-import { ACQUISITION_UNIT_POLICY_TYPE } from '../../constants';
-import { AcquisitionUnitPolicy } from './PolicyRenderComponents';
+import { ACQUISITION_UNIT_POLICY_TYPE } from '../../../constants';
+import { AcquisitionUnitPolicy } from '../PolicyRenderComponents';
 
 const PolicyTypedown = ({
   policies = [],
