+++ lib/AccessControl/Form/PolicyTypedown/PolicyTypedown/PolicyTypedown.test.js
@@ -7,9 +7,11 @@ import {
   mockTypedownGetter,
   renderWithIntl
 } from '@folio/stripes-erm-testing';
-import { PolicyTypedown } from './index';
-import { translationsProperties } from '../../../../test/jest/helpers';
-import { flattenedPolicies } from '../../../../test/jest/resources';
+
+import PolicyTypedown from './PolicyTypedown';
+
+import { translationsProperties } from '../../../../../test/jest/helpers';
+import { flattenedPolicies } from '../../../../../test/jest/resources';
 
 const POLICY_TYPEDOWN_ID = 'policy-typedown-test';
 const POLICY_TYPEDOWN_LABEL = `Typedown-${POLICY_TYPEDOWN_ID}`;
@@ -52,7 +54,7 @@ jest.mock('@k-int/stripes-kint-components', () => {
   });
 });
 
-jest.mock('./PolicyRenderComponents', () => {
+jest.mock('../PolicyRenderComponents', () => {
   return ({
     AcquisitionUnitPolicy: jest.fn(({ policy: { policy: { name } = {} } = {} }) => (
       <>
