+++ lib/AccessControl/View/AccessControl/AccessControl.js
@@ -1,6 +1,6 @@
 import { Accordion, Badge } from '@folio/stripes/components';
 import { FormattedMessage } from 'react-intl';
-import PoliciesTable from './PoliciesTable';
+import PoliciesTable from '../PoliciesTable';
 
 const AccessControl = ({
   onToggle,
