+++ lib/AccessControl/View/PoliciesTable/PoliciesTable.js
@@ -5,7 +5,7 @@ import { FormattedMessage, useIntl } from 'react-intl';
 
 import { IconButton, MultiColumnList } from '@folio/stripes/components';
 
-import { acquisitionPolicyRestrictions } from '../Form';
+import { acquisitionPolicyRestrictions } from '../../Form';
 
 const PoliciesTable = ({
   policies,
