+++ test/jest/resources/AccessControl/Policies/AcquisitionUnits/AcquisitionUnitPolicies.js
@@ -1,11 +1,3 @@
-import {
-  ACQUISITION_UNIT_MEMBER_NON_RESTRICTIVE,
-  ACQUISITION_UNIT_MEMBER_RESTRICTIVE,
-  ACQUISITION_UNIT_NON_MEMBER_NON_RESTRICTIVE,
-  ACQUISITION_UNIT_POLICY_TYPE
-} from '../../../../../../lib/AccessControl';
-import { claimFlattener } from '../../../../../../lib/AccessControl/hooks/useClaimPolicies';
-
 export const AUPolicyMain = {
   id: '0ebb1f7d-983f-3026-8a4c-5318e0ebc041',
   name: 'main',
@@ -136,72 +128,3 @@ export const AUPolicyRestrictDeleteNonMember = {
   isMember: false,
   deleted: false
 };
-
-export const readPolicies = {
-  readPolicies: [
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyRestrictReadMember
-      ],
-      name: ACQUISITION_UNIT_MEMBER_RESTRICTIVE
-    },
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyMain,
-        AUPolicyUnrestrictedNonMember,
-        AUPolicyRestrictCreateNonMember,
-        AUPolicyRestrictUpdateNonMember,
-        AUPolicyRestrictDeleteNonMember
-      ],
-      name: ACQUISITION_UNIT_NON_MEMBER_NON_RESTRICTIVE
-    },
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyUnrestrictedMember,
-        AUPolicyRestrictCreateMember,
-        AUPolicyRestrictUpdateMember,
-        AUPolicyRestrictDeleteMember
-      ],
-      name: ACQUISITION_UNIT_MEMBER_NON_RESTRICTIVE
-    }
-  ]
-};
-
-export const claimPolicies = {
-  claimPolicies: [
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyRestrictCreateMember
-      ],
-      name: ACQUISITION_UNIT_MEMBER_RESTRICTIVE
-    },
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyMain,
-        AUPolicyUnrestrictedNonMember,
-        AUPolicyRestrictReadNonMember,
-        AUPolicyRestrictUpdateNonMember,
-        AUPolicyRestrictDeleteNonMember
-      ],
-      name: ACQUISITION_UNIT_NON_MEMBER_NON_RESTRICTIVE
-    },
-    {
-      type: ACQUISITION_UNIT_POLICY_TYPE,
-      policies: [
-        AUPolicyUnrestrictedMember,
-        AUPolicyRestrictReadMember,
-        AUPolicyRestrictUpdateMember,
-        AUPolicyRestrictDeleteMember
-      ],
-      name: ACQUISITION_UNIT_MEMBER_NON_RESTRICTIVE
-    }
-  ]
-};
-
-// I don't like that I've had to do this in the test resources file too
-export const flattenedPolicies = claimFlattener(claimPolicies.claimPolicies);
